requirements.txt
matplotlib
numpy
Pillow
scikit-image